requirement.txt
requests
pyyaml
bs4
selenium
